@@ -8,12 +8,12 @@
   "license": "MIT",
   "devDependencies": {
     "babel-core": "^6.25.0",
-    "babel-loader": "6.2.10",
+    "babel-loader": "^7.1.1",
     "babel-preset-es2015": "^6.24.1",
     "bootstrap-datepicker": "^1.7.0",
     "bootstrap-sass": "~3.3",
     "bootstrap-tagsinput": "~0.7",
-    "bootstrap4": "npm:bootstrap@4.0.0-alpha.6",
+    "bootstrap4": "npm:bootstrap@4.0.0-beta",
     "cropper": "^3.0.0-rc.2",
     "del": "^3.0.0",
     "fancybox": "^3.0.0",
@@ -35,13 +35,14 @@
     "jquery-ui-dist": "~1.12",
     "lodash": "^4.17.4",
     "photoswipe": "^4.1.1",
+    "popper.js": "^1.11.1",
     "requestanimationframe": "^0.0.23",
     "simple-ajax-uploader": "^2.5.5",
     "slick-carousel": "^1.6.0",
     "standard": "^10.0.3",
     "sweet-scroll": "^2.2.1",
-    "webpack": "1.14.0",
-    "webpack-stream": "3.2.0"
+    "webpack": "^3.5.4",
+    "webpack-stream": "^4.0.0"
   },
   "babel": {
     "presets": [
